'more' is better place for Riemann Handler
