+++ logbook/more.py
@@ -10,6 +10,7 @@
 """
 import re
 import os
+import platform
 from collections import defaultdict
 from functools import partial
 
@@ -22,6 +23,7 @@ from logbook.helpers import PY2, string_types, iteritems, u
 
 from logbook.ticketing import TicketingHandler as DatabaseHandler
 from logbook.ticketing import BackendBase
+from riemann_client import client as riemann, transport
 
 if PY2:
     from urllib import urlencode
@@ -447,3 +449,52 @@ class DedupHandler(Handler):
                 dispatch = dispatch_record
             dispatch(record)
         self.clear()
+
+
+class RiemannHandler(Handler):
+
+    """
+    A handler that sends logs as events to Riemann.
+    """
+
+    def __init__(self, host, port, message_type="tcp", ttl=60, bubble=False):
+        self.host = host
+        self.port = port
+        self.ttl = ttl
+        if message_type == "tcp":
+            self.transport = transport.TCPTransport
+        elif message_type == "udp":
+            self.transport = transport.UDPTransport
+        else:
+            msg = ("Currently supported message types for RiemannHandler are: {0}. \
+                    {1} is not supported."
+                   .format(",".join(["tcp", "udp"], message_type)))
+            raise RuntimeError(msg)
+
+    def record_to_event(self, record):
+        tags = ["log", record.level_name]
+        msg = record.exc_info if record.exc_info else record.msg
+        channel_name = str(record.channel) if record.channel else "unknown"
+        if any([record.level_name == keywords
+                for keywords in ["ERROR", "EXCEPTION"]]):
+            state = "error"
+        else:
+            state = "ok"
+        return {"metric_f": 1.0,
+                "tags": tags,
+                "description": msg,
+                "time": record.time.isoformat(),
+                "ttl": self.ttl,
+                "host": platform.node(),
+                "service": "{0}.{1}".format(channel_name, os.getpid()),
+                "state": state
+                }
+
+    def emit(self, record):
+        self.emit_batch([record])
+
+    def emit_batch(self, records):
+        with riemann.QueuedClient(self.transport(self.host, self.port)) as cl:
+            for record in records:
+                cl.event(**self.record_to_event(record))
+            cl.flush()
